requirements.txt
certifi
selenium==4.2.0
openai
pandas
python-dotenv
chromedriver-autoinstaller
geckodriver-autoinstaller
urllib3
